requirements.txt
Django==1.9.5
django-factory-boy==1.0.0
factory-boy==2.6.1
fake-factory==0.5.7
Pillow==3.2.0
psycopg2==2.6.1
python-dateutil==2.5.2
six==1.10.0
